Reorganise core API module
